+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 
 ### Bug Fixes
 - Fixed a snapsync issue that can sometimes block the healing step [#3920](https://github.com/hyperledger/besu/pull/3920)
-- remove peer block height requirements for block fetching
+- remove peer block height requirements for block fetching [#3911](https://github.com/hyperledger/besu/pull/3911)
+- allow upgrade whether websockets enabled or not [#4019](https://github.com/hyperledger/besu/pull/4019)
 
 ## 22.4.3